+++ cubrid/CMakeLists.txt
@@ -268,7 +268,7 @@ set(JSP_HEADERS
 set(METHOD_SOURCES
   ${METHOD_DIR}/method_connection.cpp
   ${METHOD_DIR}/method_def.cpp
-  ${METHOD_DIR}/method_invoke.cpp
+  ${METHOD_DIR}/method_invoke_builtin.cpp
   ${METHOD_DIR}/method_invoke_group.cpp
   ${METHOD_DIR}/method_invoke_java.cpp
   ${METHOD_DIR}/method_query_cursor.cpp
